initial attempt at class based version of Transformer
