+++ sphinx/api/proj.rst
@@ -5,19 +5,8 @@ Proj
 pyproj.Proj
 -----------
 
-.. autoclass:: pyproj.Proj
+.. autoclass:: pyproj.proj.Proj
     :members:
     :inherited-members:
     :special-members: __init__, __call__
     :show-inheritance:
-
-pyproj.transform
-----------------
-
-.. autofunction:: pyproj.transform
-
-
-pyproj.itransform
------------------
-
-.. autofunction:: pyproj.itransform
