+++ sphinx/api/transformer.rst
@@ -1,0 +1,19 @@
+Transformer
+===========
+
+pyproj.Transformer
+------------------
+
+.. autoclass:: pyproj.transformer.Transformer
+    :members:
+
+pyproj.transform
+----------------
+
+.. autofunction:: pyproj.transformer.transform
+
+
+pyproj.itransform
+-----------------
+
+.. autofunction:: pyproj.transformer.itransform
